+++ bootstrap.js
@@ -1,3 +1,3 @@
+import Database from "@bejibun/database";
 import BaseModel from "./bases/BaseModel";
-import { initDatabase } from "./config/database";
-BaseModel.knex(initDatabase());
+BaseModel.knex(Database.knex());