+++ commands/InstallCommand.d.ts
@@ -1,4 +1,4 @@
-export default class MigrateFreshCommand {
+export default class InstallCommand {
     /**
      * The name and signature of the console command.
      *
@@ -20,8 +20,8 @@ export default class MigrateFreshCommand {
     /**
      * The arguments of the console command.
      *
-     * @var $arguments Array<Array<string>>
+     * @var $arguments Array<Array<any>>
      */
-    protected $arguments: Array<Array<string>>;
+    protected $arguments: Array<Array<any>>;
     handle(options: any, args: Array<string>): Promise<void>;
 }