+++ config/database.d.ts
@@ -1,4 +1,0 @@
-import type { Knex } from "knex";
-declare const config: Knex.Config;
-export declare const initDatabase: () => Knex;
-export default config;