+++ src/bootstrap.ts
@@ -1,5 +1,5 @@
+import Database from "@bejibun/database";
 import {Model} from "objection";
 import BaseModel from "@/bases/BaseModel";
-import {initDatabase} from "@/config/database";
 
-(BaseModel as any as typeof Model).knex(initDatabase());
+(BaseModel as any as typeof Model).knex(Database.knex());